Add hover menu to messages
Adds a menu, currently only consisting of a "start thread" button. It appears
on hover of a message -- except if that message already has a thread. This way
it does not double up with the "View thread" link.
In addition, on hover, messages now get a glow around them to show which
message you are over.